Split code into class-per-file Add psr-4 directive in composer.json
