+++ code/Controllers/CMSSiteTreeFilter.php
@@ -240,220 +240,3 @@ abstract class CMSSiteTreeFilter extends Object implements LeftAndMain_SearchFil
 		return $ids;
 	}
 }
-
-/**
- * This filter will display the SiteTree as a site visitor might see the site, i.e only the
- * pages that is currently published.
- *
- * Note that this does not check canView permissions that might hide pages from certain visitors
- *
- * @package cms
- * @subpackage content
- */
-class CMSSiteTreeFilter_PublishedPages extends CMSSiteTreeFilter {
-
-	/**
-	 * @return string
-	 */
-	static public function title() {
-		return _t('CMSSIteTreeFilter_PublishedPages.Title', "Published pages");
-	}
-
-	/**
-	 * @var string
-	 */
-	protected $childrenMethod = "AllHistoricalChildren";
-
-	/**
-	 * @var string
-	 */
-	protected $numChildrenMethod = 'numHistoricalChildren';
-
-	/**
-	 * Filters out all pages who's status who's status that doesn't exist on live
-	 *
-	 * @see {@link SiteTree::getStatusFlags()}
-	 * @return SS_List
-	 */
-	public function getFilteredPages() {
-		$pages = Versioned::get_including_deleted('SilverStripe\\CMS\\Model\\SiteTree');
-		$pages = $this->applyDefaultFilters($pages);
-		$pages = $pages->filterByCallback(function(SiteTree $page) {
-			return $page->getExistsOnLive();
-		});
-		return $pages;
-	}
-}
-
-/**
- * Works a bit different than the other filters:
- * Shows all pages *including* those deleted from stage and live.
- * It does not filter out pages still existing in the different stages.
- *
- * @package cms
- * @subpackage content
- */
-class CMSSiteTreeFilter_DeletedPages extends CMSSiteTreeFilter {
-
-	/**
-	 * @var string
-	 */
-	protected $childrenMethod = "AllHistoricalChildren";
-
-	/**
-	 * @var string
-	 */
-	protected $numChildrenMethod = 'numHistoricalChildren';
-
-	static public function title() {
-		return _t('CMSSiteTreeFilter_DeletedPages.Title', "All pages, including archived");
-	}
-
-	public function getFilteredPages() {
-		$pages = Versioned::get_including_deleted('SilverStripe\\CMS\\Model\\SiteTree');
-		$pages = $this->applyDefaultFilters($pages);
-		return $pages;
-	}
-}
-
-/**
- * Gets all pages which have changed on stage.
- *
- * @package cms
- * @subpackage content
- */
-class CMSSiteTreeFilter_ChangedPages extends CMSSiteTreeFilter {
-
-	static public function title() {
-		return _t('CMSSiteTreeFilter_ChangedPages.Title', "Modified pages");
-	}
-
-	public function getFilteredPages() {
-		$pages = Versioned::get_by_stage('SilverStripe\\CMS\\Model\\SiteTree', 'Stage');
-		$pages = $this->applyDefaultFilters($pages)
-			->leftJoin('SiteTree_Live', '"SiteTree_Live"."ID" = "SiteTree"."ID"')
-			->where('"SiteTree"."Version" <> "SiteTree_Live"."Version"');
-		return $pages;
-	}
-}
-
-/**
- * Filters pages which have a status "Removed from Draft".
- *
- * @package cms
- * @subpackage content
- */
-class CMSSiteTreeFilter_StatusRemovedFromDraftPages extends CMSSiteTreeFilter {
-
-	static public function title() {
-		return _t('CMSSiteTreeFilter_StatusRemovedFromDraftPages.Title', 'Live but removed from draft');
-	}
-
-	/**
-	 * Filters out all pages who's status is set to "Removed from draft".
-	 *
-	 * @return SS_List
-	 */
-	public function getFilteredPages() {
-		$pages = Versioned::get_including_deleted('SilverStripe\\CMS\\Model\\SiteTree');
-		$pages = $this->applyDefaultFilters($pages);
-		$pages = $pages->filterByCallback(function(SiteTree $page) {
-			// If page is removed from stage but not live
-			return $page->getIsDeletedFromStage() && $page->getExistsOnLive();
-		});
-		return $pages;
-	}
-}
-
-/**
- * Filters pages which have a status "Draft".
- *
- * @package cms
- * @subpackage content
- */
-class CMSSiteTreeFilter_StatusDraftPages extends CMSSiteTreeFilter {
-
-	static public function title() {
-		return _t('CMSSiteTreeFilter_StatusDraftPages.Title', 'Draft pages');
-	}
-
-	/**
-	 * Filters out all pages who's status is set to "Draft".
-	 *
-	 * @see {@link SiteTree::getStatusFlags()}
-	 * @return SS_List
-	 */
-	public function getFilteredPages() {
-		$pages = Versioned::get_by_stage('SilverStripe\\CMS\\Model\\SiteTree', 'Stage');
-		$pages = $this->applyDefaultFilters($pages);
-		$pages = $pages->filterByCallback(function(SiteTree $page) {
-			// If page exists on stage but not on live
-			return (!$page->getIsDeletedFromStage() && $page->getIsAddedToStage());
-		});
-		return $pages;
-	}
-}
-
-/**
- * Filters pages which have a status "Deleted".
- *
- * @package cms
- * @subpackage content
- */
-class CMSSiteTreeFilter_StatusDeletedPages extends CMSSiteTreeFilter {
-
-	/**
-	 * @var string
-	 */
-	protected $childrenMethod = "AllHistoricalChildren";
-
-	/**
-	 * @var string
-	 */
-	protected $numChildrenMethod = 'numHistoricalChildren';
-
-	static public function title() {
-		return _t('CMSSiteTreeFilter_StatusDeletedPages.Title', 'Archived pages');
-	}
-
-	/**
-	 * Filters out all pages who's status is set to "Deleted".
-	 *
-	 * @see {@link SiteTree::getStatusFlags()}
-	 * @return SS_List
-	 */
-	public function getFilteredPages() {
-		$pages = Versioned::get_including_deleted('SilverStripe\\CMS\\Model\\SiteTree');
-		$pages = $this->applyDefaultFilters($pages);
-
-		$pages = $pages->filterByCallback(function(SiteTree $page) {
-			// Doesn't exist on either stage or live
-			return $page->getIsDeletedFromStage() && !$page->getExistsOnLive();
-		});
-		return $pages;
-	}
-}
-
-/**
- * @package cms
- * @subpackage content
- */
-class CMSSiteTreeFilter_Search extends CMSSiteTreeFilter {
-
-	static public function title() {
-		return _t('CMSSiteTreeFilter_Search.Title', "All pages");
-	}
-
-	/**
-	 * Retun an array of maps containing the keys, 'ID' and 'ParentID' for each page to be displayed
-	 * in the search.
-	 *
-	 * @return SS_List
-	 */
-	public function getFilteredPages() {
-		// Filter default records
-		$pages = Versioned::get_by_stage('SilverStripe\\CMS\\Model\\SiteTree', 'Stage');
-		$pages = $this->applyDefaultFilters($pages);
-		return $pages;
-	}
-}
